#12 Add component StarRating
