+++ src/components/ui/RatingStar.jsx
@@ -1,0 +1,24 @@
+import { AiFillStar } from 'react-icons/ai';
+
+function RatingStar({ avg }) {
+  // Fill color for Star rating icon:
+  const fillStar = (avgRate, num) => {
+    return Math.round(avgRate) < num ? (
+      <AiFillStar className="text-t_light_gray_3" />
+    ) : (
+      <AiFillStar className="text-ac_yellow" />
+    );
+  };
+
+  return (
+    <div className="flex">
+      {fillStar(avg, 1)}
+      {fillStar(avg, 2)}
+      {fillStar(avg, 3)}
+      {fillStar(avg, 4)}
+      {fillStar(avg, 5)}
+    </div>
+  );
+}
+
+export default RatingStar;
