+++ src/components/ui/index.js
@@ -5,3 +5,4 @@ export { default as Modal } from './Modal';
 export { default as Tag } from './Tag';
 export { default as Tip } from './Tip';
 export { default as Input } from './Input';
+export { default as RatingStar } from './RatingStar';
