+++ src/features/course/components/CourseList/CourseItem.jsx
@@ -1,7 +1,8 @@
 // import component:
 import { Tag } from '../../../../components/ui';
+import { RatingStar } from '../../../../components/ui';
 // import icon:
-import { AiOutlineTeam, AiFillStar } from 'react-icons/ai';
+import { AiOutlineTeam } from 'react-icons/ai';
 import { FaCrown } from 'react-icons/fa';
 import { BiLock } from 'react-icons/bi';
 // import hook:
@@ -20,15 +21,6 @@ function CourseItem({ course, onClick }) {
       }, []),
     );
   }, []);
-
-  // Fill color for Star rating icon:
-  const fillStar = (avgRate, num) => {
-    return Math.round(avgRate) < num ? (
-      <AiFillStar className="text-t_light_gray_3" />
-    ) : (
-      <AiFillStar className="text-ac_yellow" />
-    );
-  };
 
   // Check if the Course is on the list Course user joined:
   const checkJoinCourse = (joinList = [], courseId) => {
@@ -58,13 +50,7 @@ function CourseItem({ course, onClick }) {
         <div className="mt-2 md:flex md justify-between md:mt-4">
           <div className="flex items-center gap-1 md:flex-col md:items-start">
             {/* Average rating info */}
-            <div className="flex">
-              {fillStar(course.avgRating, 1)}
-              {fillStar(course.avgRating, 2)}
-              {fillStar(course.avgRating, 3)}
-              {fillStar(course.avgRating, 4)}
-              {fillStar(course.avgRating, 5)}
-            </div>
+            <RatingStar avg={course.avgRating} />
             {/* Rating quantity info  */}
             <p>({course.qntRating} đánh giá)</p>
           </div>
